feat: separate character-card to server and client
